refactor: move dapp logic to separate file
